Move ParserUtil tests into respective model tests
