+++ src/test/java/seedu/address/model/person/fields/AddressTest.java
@@ -1,13 +1,20 @@
 package seedu.address.model.person.fields;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.parser.exceptions.ParseException;
+
 public class AddressTest {
+
+    private static final String WHITESPACE = " \t\r\n";
+    private static final String INVALID_ADDRESS = " ";
+    private static final String VALID_ADDRESS = "123 Main Street #0505";
 
     @Test
     public void constructor_null_throwsNullPointerException() {
@@ -27,6 +34,29 @@ public class AddressTest {
         assertDoesNotThrow(() -> new Address("-")); // one character
         assertDoesNotThrow(() ->
                            new Address("Leng Inc; 1234 Market St; San Francisco CA 2349879; USA")); // long address
+    }
+
+    @Test
+    public void parseAddress_null_throwsNullPointerException() {
+        assertThrows(NullPointerException.class, () -> Address.parseAddress((String) null));
+    }
+
+    @Test
+    public void parseAddress_invalidValue_throwsParseException() {
+        assertThrows(ParseException.class, () -> Address.parseAddress(INVALID_ADDRESS));
+    }
+
+    @Test
+    public void parseAddress_validValueWithoutWhitespace_returnsAddress() throws Exception {
+        Address expectedAddress = new Address(VALID_ADDRESS);
+        assertEquals(expectedAddress, Address.parseAddress(VALID_ADDRESS));
+    }
+
+    @Test
+    public void parseAddress_validValueWithWhitespace_returnsTrimmedAddress() throws Exception {
+        String addressWithWhitespace = WHITESPACE + VALID_ADDRESS + WHITESPACE;
+        Address expectedAddress = new Address(VALID_ADDRESS);
+        assertEquals(expectedAddress, Address.parseAddress(addressWithWhitespace));
     }
 
     @Test
