+++ src/test/java/seedu/address/model/person/fields/TagsTest.java
@@ -2,12 +2,21 @@ package seedu.address.model.person.fields;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static seedu.address.testutil.Assert.assertThrows;
+
+import java.util.Arrays;
+import java.util.Collections;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.tag.Tag;
 
 class TagsTest {
+
+    private static final String INVALID_TAG = "#friend";
+    private static final String VALID_TAG_1 = "friend";
+    private static final String VALID_TAG_2 = "neighbour";
 
     private static final String[] emptyStringArray = new String[0];
 
@@ -16,6 +25,29 @@ class TagsTest {
     @Test
     public void constructor_emptyArray_success() {
         assertEquals(new Tags(emptyTagArray), new Tags(emptyStringArray));
+    }
+
+    @Test
+    public void parseTags_null_throwsNullPointerException() {
+        assertThrows(NullPointerException.class, () -> Tags.parseTags(null));
+    }
+
+    @Test
+    public void parseTags_collectionWithInvalidTags_throwsParseException() {
+        assertThrows(ParseException.class, () -> Tags.parseTags(Arrays.asList(VALID_TAG_1, INVALID_TAG)));
+    }
+
+    @Test
+    public void parseTags_emptyCollection_returnsEmptyTags() throws Exception {
+        assertEquals(new Tags(new Tag[0]), Tags.parseTags(Collections.emptyList()));
+    }
+
+    @Test
+    public void parseTags_collectionWithValidTags_returnsTagSet() throws Exception {
+        Tags actualTags = Tags.parseTags(Arrays.asList(VALID_TAG_1, VALID_TAG_2));
+        Tags expectedTags = new Tags(new Tag(VALID_TAG_1), new Tag(VALID_TAG_2));
+
+        assertEquals(expectedTags, actualTags);
     }
 
     @Test
